+++ scout/commands/export/hpo.py
@@ -1,22 +1,25 @@
 import logging
 
 import click
+from flask.cli import with_appcontext
+
+from scout.server.extensions import store
 
 LOG = logging.getLogger(__name__)
 
 @click.command('hpo_genes', short_help='Export hpo gene list')
 @click.argument('hpo_term',nargs=-1)
-@click.pass_context
-def hpo_genes(context, hpo_term):
+@with_appcontext
+def hpo_genes(hpo_term):
     """Export a list of genes based on hpo terms"""
     LOG.info("Running scout export hpo_genes")
-    adapter = context.obj['adapter']
-    
+    adapter = store
+
     header = ["#Gene_id\tCount"]
 
     if not hpo_term:
         LOG.warning("Please use at least one hpo term")
-        context.abort()
+        raise click.Abort()
 
     for line in header:
         click.echo(line)