+++ scout/commands/index_command.py
@@ -1,6 +1,8 @@
 import logging
 
 import click
+from flask.cli import with_appcontext
+from scout.server.extensions import store
 
 LOG = logging.getLogger(__name__)
 
@@ -9,18 +11,18 @@ def abort_if_false(ctx, param, value):
         ctx.abort()
 
 @click.command('index', short_help='Index the database')
-@click.option('--yes', is_flag=True, 
+@click.option('--yes', is_flag=True,
     callback=abort_if_false,
     expose_value=False,
     prompt="This will delete and rebuild all indexes(if not --update). Are you sure?"
 )
 @click.option('--update', help="Update the indexes", is_flag=True)
-@click.pass_context
-def index(context, update):
+@with_appcontext
+def index(update):
     """Create indexes for the database"""
     LOG.info("Running scout index")
-    adapter = context.obj['adapter']
-    
+    adapter = store
+
     if update:
         adapter.update_indexes()
     else: