+++ scout/commands/query/__init__.py
@@ -1,1 +1,0 @@
-from .query_command import query