+++ scout/parse/case.py
@@ -90,6 +90,8 @@ def parse_case_data(config=None, ped=None, owner=None, vcf_snv=None,
     config_data['rank_model_version'] = config_data.get('rank_model_version')
     config_data['rank_score_threshold'] = config_data.get('rank_score_threshold', 0)
 
+    config_data['sv_rank_model_version'] = config_data.get('sv_rank_model_version')
+
     config_data['track'] = config_data.get('track', 'rare')
     if config_data['vcf_cancer']:
         config_data['track'] = 'cancer'
@@ -170,9 +172,16 @@ def parse_individual(sample):
                 'sex': str,
                 'phenotype': str,
                 'bam_file': str,
-                'vcf2cytosure': str,
+                'mt_bam': str,
                 'analysis_type': str,
+                'vcf2cytosure': str,
                 'capture_kits': list(str),
+
+                'tumor_type': str,
+                'tmb': str,
+                'msi': str,
+                'tumor_purity': str,
+                'tissue_type': str,
             }
 
     """
@@ -249,6 +258,10 @@ def parse_individual(sample):
     if tumor_purity:
         ind_info['tumor_purity'] = tumor_purity
 
+    tissue_type = sample.get('tissue_type')
+    if tissue_type:
+        ind_info['tissue_type'] = tissue_type
+
     return ind_info
 
 
@@ -311,6 +324,7 @@ def parse_case(config):
         'genome_build': config.get('human_genome_build'),
         'rank_model_version': config.get('rank_model_version'),
         'rank_score_threshold': config.get('rank_score_threshold', 0),
+        'sv_rank_model_version': config.get('sv_rank_model_version'),
         'analysis_date': config['analysis_date'],
         'individuals': individuals,
         'vcf_files': {
@@ -340,10 +354,10 @@ def parse_case(config):
             raise ValueError("madeline path not found: {}".format(mad_path))
         with mad_path.open('r') as in_handle:
             case_data['madeline_info'] = in_handle.read()
-    
+
     if (case_data['vcf_files']['vcf_cancer'] or case_data['vcf_files']['vcf_cancer_research']):
         case_data['track'] = 'cancer'
-    
+
     return case_data